+++ poly/src/test.rs
@@ -2,7 +2,7 @@ use crate::domain::*;
 use ark_ff::{PrimeField, UniformRand};
 use ark_std::test_rng;
 use ark_test_curves::bls12_381::{Fr, G1Projective};
-use ark_test_curves::mnt4_753::Fq as MNT6Fr;
+use ark_test_curves::bn384_small_two_adicity::Fr as BNFr;
 
 // Test multiplying various (low degree) polynomials together and
 // comparing with naive evaluations.
@@ -53,5 +53,5 @@ fn fft_composition() {
     test_fft_composition::<Fr, Fr, _, GeneralEvaluationDomain<Fr>>(rng, 10);
     test_fft_composition::<Fr, G1Projective, _, GeneralEvaluationDomain<Fr>>(rng, 10);
     // This will result in a mixed-radix domain being used.
-    test_fft_composition::<MNT6Fr, MNT6Fr, _, MixedRadixEvaluationDomain<_>>(rng, 17);
+    test_fft_composition::<BNFr, BNFr, _, MixedRadixEvaluationDomain<_>>(rng, 12);
 }